[feat] add check confirmation email page
